#90 feat: alterando styles do pages
